В пример добавлен координатор где рутовый контейнер UIWindow
