+++ CoordinatorExperiment/AppDelegate.swift
@@ -4,11 +4,11 @@ import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-  let window: UIWindow! = UIWindow()
+//  let window: UIWindow! = UIWindow()
   private lazy var appCoordinator: AppCoordinator = self.coordinator()
 
   func coordinator() -> AppCoordinator {
-    return AppCoordinator(window: window, initialRoute: .options)
+    return AppCoordinator(controller: LirikaWindow(container: LirikaWindow.RootContainer()), initialRoute: .options)
   }
 
   func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
