+++ sources/coords/navigation/NavigationCoordinator.swift
@@ -2,6 +2,21 @@
 
 import UIKit
 
+class LirikaNavigation: LirikaRootContaierType {
+  class RootContainer: UINavigationController {}
+
+  private let container: RootContainer
+  init(container: RootContainer? = nil) {
+    self.container = container ?? RootContainer()
+  }
+
+  func rootContainer() -> RootContainer {
+    return container
+  }
+}
+
+typealias NavigationRouter = Router<LirikaNavigation>
+
 class NavigationCoordinator<RouteType: Route>: Coordinator<RouteType, NavigationRouter> {
   override func generateRootViewController() -> LirikaNavigation {
     return LirikaNavigation(container: nil)
